finance_project/back/requirements.txt
# --- Core Web Framework ---
Django==5.2.9
djangorestframework==3.16.1
asgiref==3.11.0
sqlparse==0.5.4
tzdata==2025.3

# --- Authentication & Authorization ---
dj-rest-auth==7.0.1
django-allauth==65.13.1
# JWT 사용을 원할 경우 아래 패키지 추가가 필요할 수 있습니다.
# djangorestframework-simplejwt==5.3.1 

# --- Infrastructure & Security ---
django-cors-headers==4.9.0
django-environ==0.12.0
requests==2.32.5
urllib3==2.6.2

# --- Utilities ---
python-dateutil==2.9.0.post0
pytz==2025.2
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11